updated bug in program for UC deletion also

+++ src/Model/Program.cs
@@ -87,7 +87,7 @@ internal class Program {
             // delete src\surveyDatabase\ folder
             System.IO.DirectoryInfo di = new System.IO.DirectoryInfo(System.IO.Path.Combine(projectPath, "surveyDatabase"));
             if (di.Exists == false) {
-                Console.WriteLine("Old database not found.");
+                Console.WriteLine("Old surveyDatabase not found.");
                 return;
             }
             foreach (System.IO.FileInfo file in di.GetFiles()) {
@@ -98,6 +98,10 @@ internal class Program {
             }
 
             System.IO.DirectoryInfo di2 = new System.IO.DirectoryInfo(System.IO.Path.Combine(projectPath, "UserCredentials"));
+            if (di2.Exists == false) {
+                Console.WriteLine("Old UserCredentials not found.");
+                return;
+            }
             foreach (System.IO.FileInfo file in di2.GetFiles()) {
                 file.Delete();
             }
